requirements.txt
rockman
Django==3.2.8

dj-rest-auth==4.0.1
django-allauth==0.54.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

django-environ==0.9.0

django-ckeditor==6.3.2

drf-extra-fields==3.7.0
drf-yasg[validation]==1.21.7

idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3

boto3==1.24.41
django-storages==1.12.3
Pillow==9.2.0

djongo==1.3.6

# pyodbc==4.0.34
# mssql-django==1.2

argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0

gunicorn==20.1.0
whitenoise==6.2.0

pytz==2023.3